requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.12=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
arrow=1.2.2=pyhd3eb1b0_0
astroid=2.6.6=py38haa95532_0
asttokens=2.0.5=pyhd3eb1b0_0
async-generator=1.10=pypi_0
atomicwrites=1.4.0=py_0
attrs=21.4.0=pyhd3eb1b0_0
autopep8=1.6.0=pyhd3eb1b0_0
babel=2.9.1=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bcrypt=3.2.0=py38he774522_0
beautifulsoup4=4.11.1=py38haa95532_0
binaryornot=0.4.4=pyhd3eb1b0_1
black=19.10b0=py_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2022.4.26=haa95532_0
certifi=2022.5.18.1=py38haa95532_0
cffi=1.15.0=py38h2bbff1b_1
chardet=4.0.0=py38haa95532_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38haa95532_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyhd3eb1b0_0
cookiecutter=1.7.3=pyhd3eb1b0_0
cryptography=37.0.1=py38h21b164f_0
debugpy=1.5.1=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
diff-match-patch=20200713=pyhd3eb1b0_0
distlib=0.3.4=pypi_0
docutils=0.17.1=py38haa95532_1
entrypoints=0.4=py38haa95532_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.7.1=pypi_0
flake8=3.9.2=pyhd3eb1b0_0
flask=2.1.2=pypi_0
flask-sqlalchemy=2.5.1=pypi_0
freetype=2.10.4=h546665d_1
greenlet=1.1.2=pypi_0
gunicorn=20.1.0=pypi_0
h11=0.13.0=pypi_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
imagesize=1.3.0=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38haa95532_0
importlib_metadata=4.11.3=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
imutils=0.5.4=pypi_0
inflection=0.5.1=py38haa95532_0
intel-openmp=2021.4.0=haa95532_3556
intervaltree=3.1.0=pyhd3eb1b0_0
ipykernel=6.9.1=py38haa95532_0
ipython=8.3.0=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
isort=5.9.3=pyhd3eb1b0_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=py38haa95532_1
jinja2=3.0.3=pyhd3eb1b0_0
jinja2-time=0.2.0=pyhd3eb1b0_3
jpeg=9e=h2bbff1b_0
jsonschema=4.4.0=py38haa95532_0
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.10.0=py38haa95532_0
jupyterlab_pygments=0.1.2=py_0
keyring=23.4.0=py38haa95532_0
lazy-object-proxy=1.6.0=py38h2bbff1b_0
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libspatialindex=1.9.3=h6c2663c_0
libtiff=4.0.10=h21b02b4_1005
libwebp=1.2.2=h57928b3_0
libwebp-base=1.2.2=h8ffe710_1
lz4-c=1.9.3=h8ffe710_1
markupsafe=2.0.1=py38h2bbff1b_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mccabe=0.6.1=py38_1
mistune=0.8.4=py38he774522_1000
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
mypy_extensions=0.4.3=py38haa95532_1
nbclient=0.5.13=py38haa95532_0
nbconvert=6.4.4=py38haa95532_0
nbformat=5.3.0=py38haa95532_0
nest-asyncio=1.5.5=py38haa95532_0
numpy=1.22.4=pypi_0
numpy-base=1.22.3=py38hca35cd5_0
numpydoc=1.2=pyhd3eb1b0_0
opencv=4.0.1=py38h2a7c758_0
openssl=1.1.1o=h2bbff1b_0
outcome=1.1.0=pypi_0
packaging=21.3=pyhd3eb1b0_0
pandocfilters=1.5.0=pyhd3eb1b0_0
paramiko=2.8.1=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pathspec=0.7.0=py_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py38hdc2b20a_0
pip=21.2.2=py38haa95532_0
platformdirs=2.5.2=pypi_0
pluggy=1.0.0=py38haa95532_1
poyo=0.5.0=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
psutil=5.8.0=py38h2bbff1b_1
psycopg2=2.9.3=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py=1.11.0=pypi_0
py-opencv=4.0.1=py38he44ac1e_0
pycodestyle=2.7.0=pyhd3eb1b0_0
pycparser=2.21=pyhd3eb1b0_0
pydocstyle=6.1.1=pyhd3eb1b0_0
pyflakes=2.3.1=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pylint=2.9.6=py38haa95532_1
pyls-spyder=0.4.0=pyhd3eb1b0_0
pynacl=1.4.0=py38h62dcd97_1
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py38hd77b12b_6
pyrsistent=0.18.0=py38h196d8e1_0
pysocks=1.7.1=py38haa95532_0
pytesseract=0.3.10=pyhd8ed1ab_0
python=3.8.13=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.15.1=pyhd3eb1b0_0
python-lsp-black=1.0.0=pyhd3eb1b0_0
python-lsp-jsonrpc=1.0.0=pyhd3eb1b0_0
python-lsp-server=1.2.4=pyhd3eb1b0_0
python-slugify=5.0.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2021.3=pyhd3eb1b0_0
pywin32=302=py38h2bbff1b_2
pywin32-ctypes=0.2.0=py38_1000
pyyaml=6.0=py38h2bbff1b_1
pyzmq=22.3.0=py38hd77b12b_2
qdarkstyle=3.0.2=pyhd3eb1b0_0
qstylizer=0.1.10=pyhd3eb1b0_0
qt=5.9.7=vc14h73c81de_0
qtawesome=1.0.3=pyhd3eb1b0_0
qtconsole=5.3.0=pyhd3eb1b0_0
qtpy=2.0.1=pyhd3eb1b0_0
regex=2022.3.15=py38h2bbff1b_0
requests=2.27.1=pyhd3eb1b0_0
rope=0.22.0=pyhd3eb1b0_0
rtree=0.9.7=py38h2eaa2aa_1
selenium=4.2.0=pypi_0
setuptools=61.2.0=py38haa95532_0
sip=4.19.13=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.2.0=pypi_0
snowballstemmer=2.2.0=pyhd3eb1b0_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
soupsieve=2.3.1=pyhd3eb1b0_0
sphinx=4.4.0=pyhd3eb1b0_0
sphinxcontrib-applehelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-devhelp=1.0.2=pyhd3eb1b0_0
sphinxcontrib-htmlhelp=2.0.0=pyhd3eb1b0_0
sphinxcontrib-jsmath=1.0.1=pyhd3eb1b0_0
sphinxcontrib-qthelp=1.0.3=pyhd3eb1b0_0
sphinxcontrib-serializinghtml=1.1.5=pyhd3eb1b0_0
spyder=5.1.5=py38haa95532_1
spyder-kernels=2.1.3=py38haa95532_0
sqlalchemy=1.4.37=pypi_0
sqlite=3.38.3=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
testpath=0.5.0=pyhd3eb1b0_0
text-unidecode=1.3=pyhd3eb1b0_0
textdistance=4.2.1=pyhd3eb1b0_0
three-merge=0.1.1=pyhd3eb1b0_0
tinycss=0.4=pyhd3eb1b0_1002
tk=8.6.12=h8ffe710_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.1=py38h2bbff1b_0
tox=3.25.0=pypi_0
traitlets=5.1.1=pyhd3eb1b0_0
trio=0.20.0=pypi_0
trio-websocket=0.9.2=pypi_0
typed-ast=1.4.3=py38h2bbff1b_1
typing_extensions=4.1.1=pyh06a4308_0
ujson=5.1.0=py38hd77b12b_0
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.26.9=py38haa95532_0
vc=14.2=h21ff451_1
virtualenv=20.14.1=pypi_0
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.1.6=py38haa95532_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
werkzeug=2.1.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
wrapt=1.12.1=py38he774522_1
wsproto=1.1.0=pypi_0
xz=5.2.5=h62dcd97_1
yaml=0.2.5=he774522_0
yapf=0.31.0=pyhd3eb1b0_0
zipp=3.8.0=py38haa95532_0
zlib=1.2.12=h8cc25b3_2
zstd=1.4.9=h6255e5f_0
